[PWN-6914] Change defaultMinAmount on Buy screen

+++ p2p_wallet/Scenes/Main/Buy/BuyViewModel.swift
@@ -57,7 +57,7 @@ final class BuyViewModel: ObservableObject {
     private var tokenPrices: [Fiat: [TokenPriceKey: Double?]] = [:]
 
     // Defaults
-    private static let defaultMinAmount = Double(40)
+    private static let defaultMinAmount = Double(30)
     private static let defaultMaxAmount = Double(10000)
     private static let tokens: [Token] = [.usdc, .nativeSolana]
     private static let fiats: [Fiat] = available(.buyBankTransferEnabled) ? [.eur, .gbp, .usd] : [.usd]
